Corrige estilos do componente DesignacaoCargosCoad
